Corrected Security Entry

+++ BusinessLayer/BusinessObjects/GroupAccess.vb
@@ -13,7 +13,7 @@ Namespace BusinessObjects
 
         Public Property IdNo As Nullable(Of Int32)
         Public Property SecurityGroupIdNo As Nullable(Of Int16)
-        Public Property SecurityObjectIdNo As Nullable(Of Int16)
+        Public Property SecurityObjectIdNo As Nullable(Of Int32)
         Public Property SecurityObjectName As String
         Public Property Visible As Boolean
