+++ PresentationLayer/Models/Model.vb
@@ -314,7 +314,7 @@ Public Class Model
         Return Service.GetFieldType(tableName, fieldName)
     End Function
 
-    Public Function GetUserSecurity(securityObjectIdNo As Int16, securityGroupIdNo As Int16) As ArrayList _
+    Public Function GetUserSecurity(securityObjectIdNo As Int32, securityGroupIdNo As Int16) As ArrayList _
         Implements IModel.GetUserSecurity
         Return Service.GetUserSecurity(securityObjectIdNo, securityGroupIdNo)
     End Function
