+++ ServicesLayer/Services/IService.vb
@@ -84,7 +84,7 @@ Namespace Services
 
         Function GetControlSecurityIdNo(searchValue As String) As String
 
-        Function GetUserSecurity(securityObjectIdNo As Int16, securityGroupIdNo As Int16) As ArrayList
+        Function GetUserSecurity(securityObjectIdNo As Int32, securityGroupIdNo As Int16) As ArrayList
 
         Function GetUserSecurityForKey(securityObjectName As String, securityGroupIdNo As Int16) As ArrayList
